% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_pathway_enrich.R
\name{reactome_pathway_enrichment}
\alias{reactome_pathway_enrichment}
\title{reactome_pathway_enrichment}
\usage{
reactome_pathway_enrichment(gene_list, background_genes, p_val = 0.05)
}
\arguments{
\item{p_val}{}
}
\value{

}
\description{
reactome_pathway_enrichment
}
