% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_low_count_genes.R
\name{remove_low_count_genes}
\alias{remove_low_count_genes}
\title{remove_low_count_genes}
\usage{
remove_low_count_genes(expr_mat, no_genes = 10)
}
\arguments{
\item{no_genes}{}
}
\value{

}
\description{
remove_low_count_genes
}
