% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_rows_cols.R
\name{remove_na_rows_cols}
\alias{remove_na_rows_cols}
\title{remove_na_rows_cols}
\usage{
remove_na_rows_cols(df, col_perc_max = 10, row_perc_max = 10)
}
\arguments{
\item{row_perc_max}{}
}
\value{

}
\description{
remove_na_rows_cols
}
