% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lib_size_filter.R
\name{sample_lib_size_filter}
\alias{sample_lib_size_filter}
\title{sample_lib_size_filter}
\usage{
sample_lib_size_filter(expr_mat, min_lib_size = 1e+06)
}
\arguments{
\item{min_lib_size}{}
}
\value{

}
\description{
sample_lib_size_filter
}
